Mejoras en el entorno de pruebas
Resuelve un error a la hora de establecer el estado actual (primero poner la velocidad y luego establecer la dirección).

+++ robocar/ControlRemoto.cpp
@@ -12,7 +12,7 @@ void ControlRemoto::inicializar() {
   _receptorIR->enableIRIn(); // Start the IR receiver
 }
 
-Orden ControlRemoto::obtenerOrden(long int key) {
+Orden ControlRemoto::obtenerOrden(int key) {
   Orden orden = Orden::Indefinida;
   
   switch (key)
@@ -81,7 +81,7 @@ void ControlRemoto::reset() {
   _key = KEY_REPEAT;
 }
 
-bool ControlRemoto::setKey(int key) {
+bool ControlRemoto::setKey(long int key) {
     _key = key;
   }
 #endif
